Fix bug for snp id in review page
Originally, we didn't check if snp id was merged to another
one. So, it will crash if we want to parse hgvs from the
return html. Here we check if there is following string
"This snp_id was merged into". If there is this string,
we parse the new snp id instead of original one.