Changing PEDIA status to constant
We define 7 different status in PediaStatus.
And two functions normal_failed? and running?.
normal_failed? check if pedia service exit with normal exception
by preprocessing or workflow. If the service is killed by unknown
reason such as server crash or timeout (60 mins), pedia_service will
call error function and if the status is init or running instead of
normal_failed. Then we know it is killed by the server.

running? is for checking if service is in init or running state to
prevent another service for the same case.