Init Face2Gene LAB support
If the patient from Charite or UKB, F2G will transfer submitter and
lab info to us. If the lab info doesn't exists, we use default
Face2Gene LAB.

For the user accessing issue, we should implement it in the future.
Only the user in the same lab can access the patient in the same lab.

Closed #101